backend/api/routers/forecasts.py
"""Forecasting routes using Vertex AI."""
from fastapi import APIRouter, HTTPException, Depends, Query
import sys
from typing import List, Dict, Any

sys.path.append("../..")

from shared.auth import get_current_user_id
from google.cloud import aiplatform

router = APIRouter()


@router.get("")
async def get_production_forecast(
    tenant_id: str = Query(...),
    partner_id: str = Query(...),
    periods: int = Query(6, ge=1, le=12, description="Number of periods to forecast"),
    user_id: str = Depends(get_current_user_id),
) -> Dict[str, Any]:
    """
    Get ML-generated production forecast from Vertex AI.

    Note: This requires a deployed Vertex AI model endpoint.
    For development, this returns mock forecast data.
    """
    try:
        # TODO: Implement actual Vertex AI prediction
        # For now, return mock forecast data

        # Mock forecast data
        forecasts = []
        base_volume = 1000.0

        for i in range(periods):
            forecasts.append({
                "period": i + 1,
                "forecasted_volume": base_volume * (0.95 + (0.1 * (i % 3))),
                "confidence_lower": base_volume * 0.85,
                "confidence_upper": base_volume * 1.15,
            })

        return {
            "tenant_id": tenant_id,
            "partner_id": partner_id,
            "forecast_periods": periods,
            "forecasts": forecasts,
            "model": "vertex-ai-automl",
            "generated_at": "2025-10-24T12:00:00Z",
        }

    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to generate forecast: {str(e)}")
