backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Firebase
firebase-admin==6.3.0
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0

# Vertex AI
google-cloud-aiplatform==1.38.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Excel/PDF generation
openpyxl==3.1.2
reportlab==4.0.7

# HTTP client
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
ruff==0.1.6
mypy==1.7.1
